fix: there is absolutely no reason to keep new_landing_velocity as an uint32_t instead of a float

+++ include/skybrush/trajectory.h
@@ -202,7 +202,7 @@ sb_error_t sb_trajectory_replace_end_to_land_at(
     sb_trajectory_t* trajectory,
     sb_trajectory_stats_t* stats,
     sb_vector3_with_yaw_t new_landing_position,
-    uint32_t new_landing_velocity);
+    float new_landing_velocity);
 
 /* ************************************************************************* */
